moved connect and disconnect back to network_components
