+++ tensornetwork/network_operations.py
@@ -17,19 +17,16 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 import collections
-from typing import Any, Dict, List, Optional, Set, Text, Tuple, Type, Union, \
-  overload, Sequence
+from typing import Any, Dict, List, Optional, Set, Text, Tuple, Union, \
+    Sequence
 import numpy as np
-import weakref
-from abc import ABC, abstractmethod
-import h5py
 
 #pylint: disable=useless-import-alias
 import tensornetwork.config as config
+from tensornetwork.network_components import BaseNode, Node, CopyNode, Edge
 from tensornetwork.backends import backend_factory
 from tensornetwork.backends.base_backend import BaseBackend
-
-string_type = h5py.special_dtype(vlen=str)
+from tensornetwork.network_components import connect, disconnect
 Tensor = Any
 
 
@@ -670,46 +667,6 @@ def contract_parallel(edge: Edge) -> BaseNode:
   if edge.is_dangling():
     raise ValueError("Attempted to contract dangling edge: '{}'".format(edge))
   return contract_between(edge.node1, edge.node2)
-
-
-def connect(edge1: Edge, edge2: Edge, name: Optional[Text] = None) -> Edge:
-  for edge in [edge1, edge2]:
-    if not edge.is_dangling():
-      raise ValueError("Edge '{}' is not a dangling edge. "
-                       "This edge points to nodes: '{}' and '{}'".format(
-                           edge, edge.node1, edge.node2))
-  if edge1 is edge2:
-    raise ValueError("Cannot connect and edge '{}' to itself.".format(edge1))
-
-  if edge1.dimension != edge2.dimension:
-    raise ValueError("Cannot connect edges of unequal dimension. "
-                     "Dimension of edge '{}': {}, "
-                     "Dimension of edge '{}': {}.".format(
-                         edge1, edge2.dimension, edge2, edge2.dimension))
-
-  #edge1 and edg2 are always dangling in this case
-  node1 = edge1.node1
-  node2 = edge2.node1
-  axis1_num = node1.get_axis_number(edge1.axis1)
-  axis2_num = node2.get_axis_number(edge2.axis1)
-
-  new_edge = Edge(
-      node1=node1, axis1=axis1_num, name=name, node2=node2, axis2=axis2_num)
-
-  node1.add_edge(new_edge, axis1_num, override=True)
-  node2.add_edge(new_edge, axis2_num, override=True)
-  return new_edge
-
-
-def disconnect(edge, edge1_name: Optional[Text],
-               edge2_name: Optional[Text]) -> Tuple[Edge, Edge]:
-  """
-  Break an existing non-dangling edge.
-  This updates both Edge.node1 and Edge.node2 by removing the 
-  connecting edge from `Edge.node1.edges` and `Edge.node2.edges`
-  and adding new dangling edges instead
-  """
-  return edge.disconnect(edge1_name, edge2_name)
 
 
 def conj(node: BaseNode,
